backend/requirements.txt
fastapi
uvicorn
python-dotenv
PyJWT
httpx
